File hierarchy refactory
